8팀 2주차 코드리뷰 요청

See merge request sw_track/class_07/web_project_2/team8/server!38

+++ src/controllers/postController.js
@@ -1,36 +1,92 @@
+import commonError from '../constants/errorConstant.js';
+import CustomError from '../middleware/errorHandler.js';
 import * as postService from '../services/postService.js';
 
-const userId = '658147ffc84ca272c761ec03';
+export async function getPosts(req, res) {
+  const filter = req.query;
 
-// 모든 게시글 조회
-export async function getAllPosts(req, res) {
-  const posts = await postService.getAllPosts();
+  let posts = [];
 
-  if (!posts) {
-    return res.status(200).json({ message: '전체 게시글을 찾을 수 없습니다.' });
+  // 전체 조회
+  if (Object.keys(filter).length === 0 || !filter.tag) {
+    const allPosts = await postService.getAllPosts();
+    posts.push(...allPosts);
   }
 
-  res.status(200).json({ posts });
+  // 여행지로 조회
+  if (filter.city) {
+    const postsByDestination = await postService.getAllPostsByDestination(filter.city);
+    posts.push(...postsByDestination);
+  }
+
+  // 태그로 조회
+  if (filter.tag) {
+    const tags = filter.tag.split(',');
+    const postsByTags = await postService.getAllPostsByTags(tags);
+    posts.push(...postsByTags);
+  }
+
+  if (filter.sort) {
+    const sort = filter.sort;
+
+    // 정렬기준(최신순, 오래된순, 찜많은 순) 중복 선택되었는지 검사
+    if (sort.split(',').length > 1) {
+      throw new CustomError(commonError.TAG_COUNT_ERROR, '정렬 기준 선택지는 하나만 선택 가능합니다.', {
+        statusCode: 404,
+      });
+    }
+
+    const sortedPosts = await postService.getPostsBySort(sort, posts);
+    posts = sortedPosts;
+  }
+
+  return res.status(200).json({ posts });
 }
 
 // 특정 게시글 조회
-export async function getPostByPostId(req, res) {
+export async function getPostById(req, res) {
   const postId = req.params.postId;
-  const post = await postService.getPostByPostId(postId);
 
-  if (!post) {
-    return res.status(404).json({ message: '해당 게시글을 찾을 수 없습니다.' });
+  if (!postId) {
+    throw new CustomError(commonError.POST_UNKNOWN_ERROR, '해당 게시글의 고유 아이디 값이 없습니다.', {
+      statusCode: 404,
+    });
   }
 
-  return res.status(200).json({ message: '해당 게시글이 조회되었습니다.' });
+  const post = await postService.getPostById(postId);
+
+  return res.status(200).json(post);
+}
+
+// 특정 사용자의 게시글 조회
+export async function getAllPostsByUserId(req, res) {
+  const userId = req.userId;
+
+  if (!userId) {
+    throw new CustomError(commonError.USER_UNKNOWN_ERROR, '조회하려는 특정 사용자를 찾을 수 없습니다.', {
+      statusCode: 404,
+    });
+  }
+
+  const posts = await postService.getAllPostsByUserId(userId);
+
+  return res.status(200).json({ posts });
 }
 
 // 게시글 추가
 export async function createPost(req, res) {
+  const userId = req.userId;
+
+  if (!userId) {
+    throw new CustomError(commonError.USER_UNKNOWN_ERROR, '추가하려는 특정 사용자를 찾을 수 없습니다.', {
+      statusCode: 404,
+    });
+  }
+
   const { title, destination, startDate, endDate, tag, schedules, distances, cost, peopleCount, isPublic, reviewText } =
     req.body;
 
-  await postService.createPost(userId, {
+  const result = await postService.createPost(userId, {
     title,
     destination,
     startDate,
@@ -44,12 +100,36 @@ export async function createPost(req, res) {
     reviewText,
   });
 
-  res.status(201).json({ message: '게시글이 등록되었습니다.' });
+  if (!result) {
+    throw new CustomError(commonError.POST_UNKNOWN_ERROR, '이미 같은 ID로 게시글이 등록되어 있습니다.', {
+      statusCode: 404,
+    });
+  }
+
+  return res.status(201).json({ message: '게시글이 등록되었습니다.' });
 }
 
 // 특정 사용자의 게시글 수정
 export async function updatePost(req, res) {
+  const userId = req.userId;
   const postId = req.params.postId;
+
+  if (!userId) {
+    throw new CustomError(commonError.USER_UNKNOWN_ERROR, '수정하려는 특정 사용자를 찾을 수 없습니다.', {
+      statusCode: 404,
+    });
+  }
+
+  if (!postId) {
+    throw new CustomError(
+      commonError.POST_UNKNOWN_ERROR,
+      '수정하려는 해당 게시글의 고유 아이디값을 찾을 수 없습니다.',
+      {
+        statusCode: 404,
+      },
+    );
+  }
+
   const { title, destination, startDate, endDate, tag, schedules, distances, cost, peopleCount, isPublic, reviewText } =
     req.body;
 
@@ -68,20 +148,42 @@ export async function updatePost(req, res) {
   });
 
   if (!result) {
-    return res.status(404).json({ message: '해당 게시글을 찾을 수 없습니다.' });
+    throw new CustomError(commonError.POST_UNKNOWN_ERROR, '해당 게시글을 찾을 수 없습니다.', {
+      statusCode: 404,
+    });
   }
 
-  return res.status(201).json({ message: '게시글이 수정되었습니다.' });
+  return res.status(200).json({ message: '게시글이 수정되었습니다.' });
 }
 
 // 특정 사용자의 게시글 삭제
 export async function deletePost(req, res) {
+  const userId = req.userId;
   const postId = req.params.postId;
-  const post = await postService.deletePost(userId, postId);
 
-  if (post === null) {
-    return res.status(404).json({ message: '해당 게시글이 존재하지 않습니다.' });
+  if (!userId) {
+    throw new CustomError(commonError.USER_UNKNOWN_ERROR, '삭제하려는 특정 사용자를 찾을 수 없습니다.', {
+      statusCode: 404,
+    });
   }
 
-  res.status(204).json({ message: '게시글 삭제 성공' });
+  if (!postId) {
+    throw new CustomError(
+      commonError.POST_UNKNOWN_ERROR,
+      '삭제하려는 해당 게시글의 고유 아이디값을 찾을 수 없습니다.',
+      {
+        statusCode: 404,
+      },
+    );
+  }
+
+  const result = await postService.deletePost(userId, postId);
+
+  if (!result) {
+    throw new CustomError(commonError.POST_DELETE_ERROR, '해당 게시글이 존재하지 않아 삭제할 수 없습니다', {
+      statusCode: 404,
+    });
+  }
+
+  return res.status(204);
 }
